[StationeryPalette] 64bit 環境でエラーが発生する不具合を修正。

+++ FileTreeView/FileTreeNode.m
@@ -474,7 +474,7 @@ NSString *ORDER_CHACHE_NAME = @"order.plist";
 
 - (FileTreeNode *)createFolderAfter:(FileTreeNode *)item withName:(NSString *)aName
 {
-	int child_index;
+	NSInteger child_index;
 	if (item) {
 		child_index = [self indexOfChild:item];
 	}
