+++ FileTreeView/TreeNode.m
@@ -100,7 +100,7 @@
 }
 
 - (void)removeChild:(TreeNode*)child {
-    int index = [self indexOfChild: child];
+    NSInteger index = [self indexOfChild: child];
     if (index!=NSNotFound) {
         [self _removeChildrenIdenticalTo: [NSArray arrayWithObject: [self childAtIndex:index]]];
     }
